Raise error if UploadedFile received invalid data
It can happen that someone sends a JSON representation of the uploaded
file with a key missing or misspelled, or initializes `{}` as a default
column value. In these cases Shrine::UploadedFile would raise a KeyError
for the hash key that is missing.

That is not a very self-explanatory error, so in this commit we add a
check for whether the data hash contains the needed keys, and if it
doesn't raises an explicit Shrine::Error that this is not valid data.

Before:

  Shrine::UploadedFile.new({}) # KeyError: key not found: "storage"

After:

  Shrine::UploadedFile.new({}) # Shrine::Error: {} isn't valid uploaded file data