Support plain models in backgrounding plugin
Previously the backgrounding plugin only worked for model instances that
represent database records. It relied on `Attacher.find_record` being
defined, which comes with ORM plugins.

However, Shrine can be also used with plain models, and there is no real
reason not to support that in backgrounding plugin as well. It only wouldn't
be possible if Shrine was obtaining the cached file from the database,
but the cached file is sent in the background job arguments, so it's
available regardless of whether it's persisted or not.

This commit adds support for plain models (if `Attacher.find_record`
isn't defined), and it is used the same way as with database records.

  Shrine::Attacher.promote { |data| PromoteJob.perform_async(data) }

  # ...

  class PromoteJob
    def perform(data)
      attacher = Shrine::Attacher.promote(data)
      attacher.get # returns promoted files
    end
  end

The returned attacher object can then be used to get the promoted
files, so that you can perform any additional tasks if you need to.