Detect when two versions point to same IO object
Sometimes it can happen that, due to a mistake in writing the processing
code, a hash of versions is sent where two versions point to the same IO
object. An example of such processing code:

  process(:store) do |io|
    original = io.download
    resized = resize_to_limit!(original, 800, 800)
    {original: original, resized: resized}
  end

Since the destructive resizing was done instead of nondestructive,
`original` and `resize` are now the same Tempfile object. Now when
Shrine proceeds to upload these, uploading the first version will
close the Tempfile, and then the second version will fail to upload,
because the file has already been closed.

Instead of letting an IOError be raised, this commit adds a detection
before uploading, and raises a descriptive Shrine::Error showing which
versions were pointing to the same IO object.