Add Attacher.default_url for declaring default URL
Previously the way to declare the default URL was

  plugin :default_url do |context|
    # ...
  end

I find this a bit awkward, and it makes it seem like this default URL is
used on the uploader level, when in fact this behaviour is tied to the
attacher only. This commit adds `Attacher.default_url` which has the
same behaviour:

  plugin :default_url

  Attacher.default_url do |options|
    # ...
  end

Having attacher class methods already exists in Shrine; we have
`Attacher.promote` and `Attacher.delete` from backgrounding plugin, and
`Attacher.validate` for declaring validations.

Since it doesn't make sense to have two ways of declaring default URLs,
we deprecate the old way of passing a block when loading the plugin.