Allow uploaders to accept Rack uploaded files
Previously loading the `rack_file` plugin would enable the attacher to
accept Rack uploaded file hashes, but not the uploaders themselves. This
is limiting for people that want to use uploaders directly, without the
attacher interface.

Since there is no real reason to have this limitation, we move accepting
uploaded files to uploaders themselves.

  params[:file] #=> {:tempfile => #<Tempfile...>, :name => "file", ...}
  uploader.upload(params[:file])

That way both the attacher and the uploader can accept Rack uploaded
file hashes.

While we're here, we also make the check for whether the hash represents a
Rack uploaded file stricter, requiring that the `:name` key is also
present.