Raise error when #generate_location returns nil
Since user can override `Shrine#generate_location`, it can happen that
due to a mistake in writing the method it returns nil for a file. In
that case most storages will eventually raise an error, because they
need a location.

However, this error will be different for every storage, and likely be
obscure. It would be better if Shrine raised an explicit error when
`Shrine#generate_location` returned nil. This is what this commit does.

Fixes #111