Rename Attacher#attached? to Attacher#changed?
The Attacher#attached? method returns whether the attachment has
changed, in other words when the attacher is "dirty". However, that
means "attached" isn't the right name for it, because I use that term
whenever I want to say the opposite of "no file isn't attached".

So we rename it to the more accurate Attacher#changed?, and keep the
Attacher#attached? alias for backwards compatibility, probably to be
deprecated in the future.