Don't dirty attacher in Attacher#set when attachment hasn't changed
Previously this logic was in Attacher#assign, because historically the
way that you would handle retaining cached file in case of validation
errors was to assign `#<attachment>_data` to the form field. If this
logic was in Attacher#set then, the `restore_cached_data` plugin would
trigger whenever a form is saved with the cached attachment (while the
background job is working).

However, now `cached_attachment_data` plugin is the idiomatic and
recommended way to retain cached file in case of validation errors,
which will assign a cached file only when it has changed. That means we
can move this logic in Attacher#set, because we want this logic in
general regardless of whether we're using Attacher#assign or
Attacher#set.