Remove only new attachments in remove_invalid plugin
The purpose of the `remove_invalid` plugin is to remove new files that
have been uploaded by the user if they turn out to be invalid. If
validations are run again later when the record has already been saved
(and the attacher is not dirty), the `remove_invalid` plugin would
currently remove the attachment.

However, it might happen that the developer runs these validations
accidentally (by excluding Attacher#changed? check), and the validation
now fails. In this case it's not correct to remove the attachment,
because that was a developer mistake; the correct thing is to fix the
validations.

This also fixes an error being raised in case attacher isn't dirty when
attempting to delete a nonexisting @old instance variable. And also it
now works when assigning stored files directly.