Exclude blacklisted types/extensions from message
Blacklists are normally used for uploads of files that would be
malicious to the storage. In this case we don't want to tell the user
which types are forbidden, because they can then use this knowledge to
easily come up with a malicious type that we had forgotten to forbid.