Deprecate regexes in format whitelists/blacklists
Shrine currently supports regexes in type/extension
whitelists/blacklists in `validation_helpers` plugin:

  validate_extension_inclusion [/jpe?g/, "png", "gif"]

These regexes are problematic because of two reasons. First and primary
reason is that these regexes need to be properly anchored and case
insensitive, so the right way to currently write the above regex is

  /\Ajpe?g\z/i

Which becomes very unreadable. The problem is that it's hard to
introduce automatic anchoring at this point without breaking backwards
compatibility.

The second reason is that it looks very ugly in error messages:

  validate_extension_inclusion [/jpe?g/, "png", "gif"]
  # "is not of allowed format (allowed formats: (?-mix:jpe?g), png, gif)"

But there is actually no good reason to keep these regexes, it's not
less convenient at all to just use strings. So we deprecate using regexes.