Deprecate dimensions validations passing when dimension is nil
If someone uploads a malformed image for which extracting dimensions
fails and remains nil, we don't want the validation to pass in this
case.

With the recent addition of conditional validations, it's easy to check
whether the image is of expected type, and only in that case perform
dimension validations.

So we now deprecate this behaviour, and in Shrine 3 the validation will
fail in case a dimension is nil.