Require Tempfile standard library in lib/shrine.rb
We are using Tempfile in `UploadedFile#download` in lib/shrine.rb, but
we weren't requiring the "tempfile" standard library. This means Shrine
would fail with a "NameError: unitialized constant Tempfile" in case
Tempfile wasn't required by other pieces of the code. This couldn't
happen inside any Rack-based web frameworks, because Rack requires
"tempfile", but it happened to me when I was writing a self-contained
script for reproducing a bug.