Add Storage::S3#client, and deprecate #s3
We currently expose Storage::S3#s3 as the only way to retrieved the
configured Aws instance, in the form of an Aws::S3::Resource object.
However, Aws::S3::Resource is a basically just a helper class to make it
more convenient to construct a bucket object and manipulate buckets, you
cannot get other useful information directly from it.

The main Aws object is Aws::S3::Client, and you can currently retrieve
it through Aws::S3::Resource#client, so in total that's
`storage.s3.client`. That's not ideal, the most useful object should be
available directly though the storage object.

So we create a new method Storage::S3#client, and deprecate
Storage::S3#s3 because it's not useful to the user, and to avoid
confusion of having two methods that access similar objects.

While we're at it, we also add tests for other methods on the storage.