Extract converting data URI into IO into Shrine.data_uri
So far we were both converting a data URI into an IO object and
assigning it to the model inside `Attacher#data_uri=`. However, one user
has requested (and rightfully so) the ability to only convert the data
URI into an IO object without assigning it, so we create a new
`Shrine.data_uri` method for that.

  Shrine.data_uri("data:image/png;base64,9kfg07a80ihtiort87t85")
  #=> #<Shrine::Plugins::DataUri::DataFile>

When parsing of the data URI fails, a
Shrine::Plugins::DataUri::ParseError is raised. This extraction also
makes it easier for us to test parsing.

While here, we also update the documentation with additional notes about
the parser, mentioning that the data URI is optional and that it doesn't
have to be base64-encoded.