Improve data URI parsing performance by switching to StringScanner
Performing a regex match on the whole data URI can be fairly slow, and
it is unnecessary considering that most of the data URI string is just
file content which doesn't need to be parsed.

So we switch to using StringScanner, to match only the amount of content
we need. In my benchmarks this has improved parsing performance 10x.