Save DataUri::DataFile to disk instead of memory
There is no reason to keep the parsed data URI in memory, so we save the
content to disk and deallocate any intermediary content strings.

This also has the added benefit that it will be detected by the S3
storage as a file, so if it's larger than 15MB aws-sdk will upload it
with parallelized multipart uploads, which can improve the overall
upload.

This change is backwards incompatible for folks who were relying on
Shrine::Plugins::DataUri::DataFile to be a subclass of StringIO.