Fix wrong implementation of FileSystem#clear!
FileSystem#clear! with an :older_than parameter was entirely skipping
directories which were modified since :older_than. However, that means
that if there are any new files in the same directory with old files,
the old files would not get deleted because the entire directory would
be skipped. Furthermore, directory's modified time also gets bumped up
when a file inside the directory is deleted, so in that case the
directory would also be skipped.

This commit simply traverses over each files (skipping directories),
deletes them if they're older than :older_than, and deletes the
underlying directories if they're empty after the deletion.