Recommend auto-orienting rotated images
Most image viewing desktop software automatically orient rotated images
(a photo can be rotated when a person taking the photo rotated their
camera by 90 degrees), so when a user is uploading a rotated image, they
might be surprised to see that the web application is displaying their
image with the wrong orientation. Therefore we want to encourage
auto-orienting uploaded images.

I thought about making auto-orientation automatically triggered by
image_processing gem in its `resize_*` methods, but each thumbnail is
being generated from the previous, we would be calling `-auto-orient`
multiple times for the image that's already auto-oriented. In my
benchmarks that didn't have any performance penalty, though, but I still
prefer that this is explicit, so that people are aware that the cached
image might not be properly oriented. I didn't want to add an example of
auto-orienting in `process(:cache)`, because that won't be triggered on
direct S3 uploads.