Enable multipart upload for FileSystem objects
When FileSystem storage is used as the temporary storage, since its
Shrine::UploadedFile won't respond to #path, it will get uploaded using
regular upload. However, if we know that the Shrine::UploadedFile is
from FileSystem storage, we can generate a direct path to the file and
give that to aws-sdk, so that it can use multipart upload when the
filesize reaches the threshold.