Accept hash with symbol keys in add_metadata block
One user tried to return a hash of metadata with symbol keys in the
`add_metadata` block, and it made things not working properly. Even
though it generally makes more sense uses string keys, because that's
how metadata is stored, the user might forget that and use symbol keys,
which is understandable considering that the singular version of
`add_metadata` also accepts a symbol argument. Since it doesn't come at
any cost to us, we make it so that any symbol keys are automatically
converted to string keys.