Deprecate passing Rack file hash to Shrine#upload
With the upcoming `Shrine.rack_file` method users will be able to
convert Rack uploaded file hashes to IO objects, and pass that to
`Shrine#upload`. This means that we can deprecate support for Rack file
hashes in `Shrine#upload`, to keep the uploader as simple as possible,
and continue supporting it only in the attacher where it's actually
convenient.