Fix download_endpoint returning incorrect responses in some cases
Currently we are streaming the response body in chunks, and we're using
a buffer string so that, instead of each new chunk allocating a new
string object, that each new chunk replaces the previous chunk in-place.
However, it seems that there are components in the Rack stack that rely
on the yielded chunks to all coexist in memory at the same time, for
example the Rack::ContentLength middleware.

Therefore we remove the buffer string and let each chunk be a new
string. Even though this means we'll be allocating more string objects
when streaming content, we need to make sure we always return the
correct content.