Fix FactoryGirl example in testing guide
If we declare the File object as a static attribute value, the same File
object will be used for each record creation, which means the second
creation will fail because Shrine will have closed the file after the
previous upload.

Instead we need to use dynamic FactoryGirl attributes so that we create
a new File object on for each new record creation.