Raise ArgumentError when :bucket is nil
Some people accidentally pass the :bucket option as nil in
Shrine::Storage::S3#initialize, for example when retrieving the value
from an environment variable which is not set.

Aws::S3::Resource#bucket doesn't tell you that you've given it nil,
instead if returns nil. Instead of leaving @bucket as nil and letting an
"undefined method on nil:NilClass" occur later on, we fail early with an
ArgumentError.