Eliminate the class variable given to the model
We don't need to assign the attacher class to a class variable on the
model, we can find the last included attachment module with the same
name, and read `#shrine_class` from it.

This reduces the pollution since it eliminates a Shrine-specific
internal class variable that the model would receive before.

It also fixes an edge case where someone might want to override the
attachment in a model subclass with a different uploader, in which case
the subclass would modify the superclass' attacher class as well (since
they share the same class variable).

class Model
  include ImageUploader::Attachment.new(:image)
end

class SpecificModel < Model
  include SpecificImageUploader::Attachment.new(:image)
end