Always require "down" in remote_url plugin
With the new HTTP.rb backend and the upcoming wget backend, there isn't
really a reason to ever use anything else than Down for downloading, so
we require it on top of the remote_url plugin file.

This removes the confusion where you would provide a custom downloader,
and be greeted with a "NameError: undefined constant Down", because
"down" would be lazily required only by the default downloader.