Utilize uploaded file metadata in download_endpoint
This modifies the download_endpoint to serialize the whole uploaded file
data in the URL, in order to be able to utilize that metadata when
generating the response. This improves the accuracy of the response in
some cases:

* The "filename" metadata value is now part of Content-Disposition
  value if present.

* If the file extension isn't registered in Rack, previously the
  Content-Type would be a generic "application/octet-stream". Now
  Content-Type will always return the value of the "mime_type" metadata
  field, which improves consistency.

* If `Storage#open` happens to return an IO object which has a blank
  #size, the endpoint will now fall back to the "size" metadata value if
  present.

The generated URLs don't have the plain storage identifier as part of
the URL anymore, it's now base64-serialized with the rest of the
uploaded file data. The new URL format can introduce a backwards
incompatibility for users relying on uploaded file ID being part of the
URL, but it's not good practice to rely on how download_endpoint
generates URLs anyway.

We select only serialize metadata values the endpoint will actually use, to
keep the URL as short as possible.

We also maintain backwards compatibility with the previous URL format,
which will probably remain supported indefinitely.