Extract rack_response plugin from download_endpoint
If a developer wants to authenticate downloads, it's difficult for them
to do it with the `download_endpoint` plugin, it's more convenient if
they can render the uploaded file from inside their router/controller.
With the rack_response plugin it's now possible to convert any
Shrine::UploadedFile object into a Rack response triple and render it.
Rails example:

  class FilesController < ActionController::Base
    def download
      # ...
      file_response = record.attachment.to_rack_response

      response.status = file_response[0]
      response.headers.merge!(file_response[1])
      self.response_body = file_response[2]
    end
  end