Split direct_upload plugin into upload_endpoint and presign_endpoint
The direct_upload plugin consists of two separate endpoints: upload
endpoint and presign endpoint. There are several downsides of this
approach:

* the /:storage/presign and /:storage/upload URL bits are hardcoded
* one endpoint cannot be mounted without the other
* user always has to load code for both features

To address these issues, we split the direct_upload plugin into
upload_endpoint and presign_endpoint plugins. The main change in these
endpoints is that they now handle root requests, which means the user can
mount them to whichever endpoint they want.

  Shrine.plugin :upload_endpoint

  Rails.application.routes.draw do
    mount Shrine.upload_endpoint(:cache) => "/images/upload"
  end

With this extraction we also remove Roda as a dependency and rewrite the
endpoints as pure Rack applications. Having Roda as a dependency wasn't
really justified, because we didn't use many of Roda's features. The
download_endpoint plugin will still depend on Roda until the next major
version, because some people might be relying on it.

The `:action` context parameter of upload_endpoint is modified from
`:cache` to `:upload`, to allow users to differentiate between
asynchronous caching and caching only after form submit. It is also
misleading if people want to use the upload endpoint for uploading to
another storage.

More things are customizable now, including the upload context,
uploading, presigning, and Rack response. It's also possible to override
these options per endpoint.

The "extension" query parameter for presign endpoint is replaced with
"filename", so that it's not ambiguous whether the extension should be
passed in with or without the leading dot. This also allows the user to
use the same parameter to set the `:content_disposition` presign option.

We also allow presigns to implement #headers and return it in the
response, because some storage services require upload requests to
include some request headers as well (Microsoft Azure Storage).

The direct_upload plugin will still remain for backwards compatibility,
but users should consider it deprecated. I will add a deprecation
warning in the next major version, just so that people can follow the
existing screencasts and blog posts without deprecation warnings.