Allow passing a block to Storage::S3#clear!
The Shrine::Storage::S3#clear! method now accepts a block, which yields
each S3 object in the storage, and deletes all objects for which the
block evaluates to true. For example, the following would delete all
objects that were uploaded more than 7 days ago:

  s3.clear! { |object| object.last_modified < Time.now - 7*24*60*60 }

This is useful when the developer doesn't have access to the S3 bucket
dashboard, and cannot configure it to automatically delete old files
from the cache.