Fix json logging failing with ActiveSupport loaded
When ActiveSupport is loaded, it monkey-patches #to_json on each object
and replaces it with its own implementation. It seems that
ActiveSupport's implementation traverses through object's instance
variables and tries to convert everything to JSON. This was throwing
errors when uploader class was supposed to get serialized, because
ActiveSupport traversed through all the #opts, and at some point it
tried to convert the logger IO (STDOUT by default) into JSON format.

We avoid that by converting both the record class and the uploader class
into strings before giving the hash to formatters.

Fixes #186