Expand relative directory path in FileSystem storage
In Rails daemon mode (`rails server -d`) the PWD (current working
directory) changes from the Rails app directory to the root directory.
In initialization phase PWD is still the Rails app directory, but as
soon as the Puma workers are spawned, the workers' PWD changes to the
root directory.

  # on initialization
  File.expand_path("public") #=> "/Users/janko/Code/rails_app/public"

  # on serving requests
  File.expand_path("public") #=> "/public"

When the FileSystem storage is instantiated with a relative directory
path as documented:

  Shrine::Storage::FileSystem.new("public", prefix: "uploads")

And the user uploads a file using this uploader, the storage will fail
with the following error:

  Errno::EACCES (Permission denied @ dir_s_mkdir - public)

This is because FileSystem storage tries to first create the directory
to the file before creating the actual file, which happens to be
`/public/...`, and this operation fails due to invalid permissions.

We fix that by expanding the path of the given directory when
initializing the FileSystem storage, so that the directory path is
always absolute.

Fixes #188