Allow changing default storage from the model instance
Previously if you would want to change the temporary/permanent storage
of the Shrine::Attacher, you would either have to use the
default_storage plugin, pass options to Shrine::Attacher.new, or use
the Shrine::Attacher directly.

The default_storage plugin is fairly limiting, because you can only
choose based on the record and attachment name. Using
`Shrine::Attacher.new` is also limiting for the same reasons. However,
while using the Shrine::Attacher directly has all the flexibility, it
can still be cumbersome to use if we're used to the attachment
interface.

We add the third option of passing attacher options via the
`#<attachment>_attacher` method:

  photo.image_attacher(cache: :other_cache, store: :other_store)
  photo.image = file # caches to :other_cache
  photo.save         # promotes to :other_store

This gives us the best of both worlds; we can still override the
temporary/permanent storages dynamically, but we don't have to repeat a
lot of boilerplate.