Deprecate the multi_delete plugin
The multi_delete plugin isn't really useful, and it adds complexity. If
users are processing and uploading multiple versions and them, and that is
too slow for them, then they'll likely want to use backgrounding and/or
parallelize plugins to speed it up. Both of these plugins can also be
used to speed up deleting, so there isn't really any value to the
multi_delete plugin.

This brings backwards incompatibility to users who are relying on
Shrine#delete accepting an array of uploaded files when loading the
versions plugin, because the versions plugin doesn't load multi_delete
anymore.