Load signature plugin globally in the docs
At the moment the documentation makes it appear as though the
`Shrine.calculate_signature` method will be available regardless of
whether the plugin is loaded globally or per uploader, which is not
true. To simplify we load the plugin globally, since there is no real
reason to load it per uploader.

Closes #215