Apply default storage options in backgrounding
When default storage options are changed in Attachment.new

  class Photo < Sequel::Model
    include Shrine::Attachment.new(:image, cache: :other_cache, store: :other_store)
  end

The backgrounding plugin wouldn't detect these options because it would
load the Attacher object without looking at the record, so it would
still promote to :store instead of :other_store.

This commit modifies the loading the Attacher to first try to load it
from the record, and only then from the `shrine_class` parameter (this
is the case when Shrine::Attacher is used standalone without
Shrine::Attachment).

Fixes #234