Rewrite demo to use Uppy
Uppy is a modern flexible non-invasive file upload library which has
support for everything that we need - uploads to the app, uploads to S3,
progress bars - so we replace the raw axios with Uppy which is
specialized in file uploads.

Uppy can also be used for uploads to Transloadit and resumable uploads
in general, so this is probably the file upload library I've been
looking for all this time.