Recommend using Uppy on the client side
* jQuery-File-Upload

Uppy is the only file upload library that integrates naturally with
Shrine, because it makes very few assumptions about the server side. It
can upload to the app, S3, and even to a tus server. I expect Uppy to
become the de-facto standard for file uploads in the browser in the near
future.