Change Logger to be shared amongst Shrine uploaders
Previously each instance of a Shrine uploader had its own instance of Logger. The logger level was set on subclassing but each uploader can override it's own logger level.

With this change, the Logger is now shared amongst all Shrine uploaders (subclasses) as there was no good reasons for duplicating Logger in uploaders. If the main Shrine class does not have a Logger instantiated or passed in, a new instance of Logger is created in the base class and then passed to the subclass. If the logger level is changed in any uploader, the change will be seen in other uploaders since the Logger is now a shared instance.