Prefer registered MIME types with the mime-types analyzer
This was specifically written to make CSV file extensions map to
text/csv instead of text/comma-separated-values when using the
mime-types gem, but will more generally prefer registered types over
un-registered types.

The IANA Media Types registry (referencing RFC 4180 and RFC 7111) say
that the media type for CSV is text/csv. Because of the way the
mime-types gem implements MIME::Types.of(filename), you get
text/comma-separated-values as listed first for a CSV:

  MIME::Types.of("csv")
  # => [#<MIME::Type: text/comma-separated-values>, #<MIME::Type: text/csv>]

Since `MIME::Type#registered?` indicates that `text/csv` is a registered
type, this change makes that the preferred media type for CSV files.

See mime-types/ruby-mime-types GitHub issue 92 for a related issue about
fixing this in the mime-types gem.