Add :mini_magick analyzer to store_dimensions plugin
ImageMagick seems to have fixed its vulnerability to image bombs in
newer versions, so it should be safe to use it for extracting
dimensions.

This is useful as folks are likely already using MiniMagick as a
dependency for processing, so this way they don't need to add an
additional dependency.

This will probably become the default analyzer in Shrine 3.