Port ImageProcessing examples to new chainable API
The new chainable API makes image_processing much more flexible and
transparent to use, and the old API has become deprecated and will be
removed in the next major version of image_processing.

This is just an intermediary phase, the processing/versions plugins will
be improved to allow automatically downloading the cached file and
deleting it after processing, which will simplify the code again. But
it's important that we do this step first.