Allow UploadedFile#download to take a block
This is useful when we want to temporarily download the file to disk,
and then have the file automatically deleted. The most common use case
would be image processing. Another use case would be checking if the
image is corrupted in file validations.

This is symmetric to UploadedFile#open which also has this feature.