Remove :rack_mime extension inferrer
The extensions in Rack::Mime::MIME_TYPES are just listed in alphabetical
order, they are not ordered according to preference. This means that
some MIME types will not get the appropriate file extension, for
example, "text/plain" will be resolved into ".conf" instead of the more
appropriate ".txt".

Therefore we cannot have :rack_mime as a supported extension inferrer,
let alone the default one. This is a backwards incompatible change for
anyone that doesn't already have the mime-types gem in their Gemfile.