Add UploadedFile#stream for streaming content into writable object
This can be useful when the user wants to write file content into an
existing writable object.

We also made it smarter; if the uploaded file is already opened, we just
stream contents and rewind the file, otherwise we open the file and then
close it after streaming. This will be used in the upcoming commit.