Use UploadedFile#stream in UploadedFile#download
The new UploadedFile#stream has the feature that when the uploaded file
is already opened, the content is streamed from the existing IO object
and the IO object is simply rewinded afterwards. This is useful when
wanting to download the uploaded file to a file while extracting
metadata, when :refresh_metadata plugin is used, as it avoids opening
the same uploaded file multiple times.