Work around a bug in older versions of file command with empty files
The file command version 5.14 has a bug where if empty content is given
on standard input, it will fail with a "no read permission". This bug
report contains more details.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=508963

This bug has since been fixed, but Ubuntu 14.04 locks the file command
version to 5.14, so anyone wanting to use a newer version of the file
command would also need to upgrade their distro.

Since it's easy enough for us to work around that bug, we modify the
code to handle that case.

Concerns #265