Don't remove bucket name from S3 URL path with :host when :force_path_style is set
The reason we currently strip away the bucket name from the path is
because when bucket name has dots in it, the URL will automatically be
generate in "path style" by aws-sdk-s3, as AWS S3 doesn't support
regular "bucket name is host" style URLs, but CDNs handle that
automatically so the CDN URLs shouldn't include the bucket name.

However, when :force_path_style is set, the user would expected the S3
URL to always be in the "path style" form, even when :host is used.

Closes #268