Don't raise exception when failing to extract dimensions
All metadata are extracted regardless of the type of file, which means
that dimensions analyzers might be called with non-image files.

In this case it's not useful to raise an exception, because it's
expected that the user might not upload a valid image. Also, the developer
might want to use the same uploader for both image and non-image files,
and extract dimensions just for image files.

So we rescue processor errors that happen during dimensions extraction
with :mini_magick and :ruby_vips analyzers, and just return nil instead.
The :fastimage analyzer already has this behaviour.