Accept Hash presigns in presign_endpoint
Having Storage#presign return an object that responds to #url and #fields is
only convenient for the S3 storage, where we actually have an
Aws::S3::PresignedPost object that responds to those methods. All other
storages need to use construct such objects using OpenStruct or Struct.
This is not nice.

This commit modifies presign_endpoint to also accept storages that
return presigns as a simple Hash, which is the first step towards
migrating to Hash presigns.