Deprecate Storage#presign returning custom object in presign_endpoint
Now that presign_endpoint supports Storage#presign returning a Hash
object, we deprecate returning a custom object that responds to #url,
\#fields and #headers, as it's now extra. Instead of strictly requiring
a Hash, we accept any object that responds to #to_h, to allow retaining
backwards compatibility in S3 storage by swapping in an object that responds
to both #url and #fields and #to_h, and also to ease off the transition,
as most storages return either Struct or OpenStruct which both implement
\#to_h.

We also add testing the #presign to the storage linter.

This introduces a backwards incompatible change for anyone relying on
S3#presign returning a Aws::S3::PresignedPost object. I think it's ok to
make this change, as it was never documented that this object is
returned in S3#presign, so users shouldn't expect it.