Extend S3#presign to support generating request data for PUT uploads
AWS SDK has two methods of generating presigned data for direct uploads.
One method is using #presigned_post, which generates request data for a
POST upload request. The other method is using #presigned_url to
generate request data for a PUT upload.

One advantage of the second method is that it allows specifying an MD5
checksum, which is something the first one doesn't support. Also, it's
generally easier to construct a PUT request where the body equals to
file content, compared to a POST request where the body needs to be
multipart-encoded.

In order to support the PUT request data, we extend S3#presign to accept
an additional ":method" option that can be set to ":put". For backwards
compatibility and the fact that it can be used in HTML forms we'll keep
POST presigns default, but we'll likely recommend using PUT method when
JavaScript is used.