Add :method key to Storage#presign result
Some storages generate presign data for POST uploads (AWS S3, Cloudinary,
Uploadcare), and other for PUT uploads (AWS S3, Google Cloud Storage,
Microsoft Azure Storage). At the moment the client cannot know whether
it should make a POST upload or a PUT upload, it has to be hardcoded,
which means it cannot be agnostic to which storage is being used.

Therefore we add an additional :method key to the Storage#presign result
which indicates which HTTP verb should be used. The name ":method" was
chosen because that parameter is already recognized by Uppy.

We also update the storage linter to require the :method key in the
presign result.