Convert all non-Ruby code blocks to Ruby
RDoc only has syntax highlighting for Ruby, so we convert all examples
to be written in pure Ruby. I think that's also nicer because we're not
assuming the ERB templating engine anymore.

The Form gem is a generic alternative to Rails form builder and SimpleForm,
so we use it for a primary examples.