Add :redirect option to download_endpoint plugin, and deprecate :storages
The :redirect option makes the download endpoint to redirect to the
direct URL of the uploaded file. This allows one to use the download
endpoint just as a means for download requests to go through the
application, which allows for authenticating download requests.

We also deprecate :storages in favour of the new
`UploadedFile#download_url` method, because with that option modifying
UploadedFile#url it wasn't possible to implement the :redirect option
without calling `Storage#url` directly. It also doesn't feel right for
UploadedFile#url to be modified to point somewhere else.